browser_service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
browser-use>=0.1.40
langchain-ollama>=0.2.0
playwright>=1.40.0
pydantic>=2.5.0
python-dotenv>=1.0.0
